+++ HISTORY.md
@@ -21,6 +21,7 @@
 - Remove unnecessary "quant" subdirectory for Salmon runs. This allows MultiQC to properly name the samples.
 - Ensure STAR log file is propagated to the upload directory.
 - Fix issue with memory not being specified properly when running `bcbio_prepare_samples.py`.
+- Run tximport automatically and store TPM in `project/date/tpm` and counts in `project/date/counts`.
 
 ## 1.1.9 (5 December 2019)
 - Fix for get VEP cache.